+++ tiledesk-client/CHANGELOG.md
@@ -1,4 +1,8 @@
-# Tiledesk Client Node.js libraries
+# Client library for Tiledesk APIs, node.js 
+
+### 0.6.19
+- added: fireEvent() raw method
+- added: fireEvent() instance method
 
 ### 0.6.18
 - added: openNow() instance method